+++ src/main/java/string/InputView.java
@@ -1,4 +1,4 @@
-package lotto;
+package string;
 
 import java.util.Scanner;
