+++ src/main/java/string/ResultView.java
@@ -1,4 +1,4 @@
-package lotto;
+package string;
 
 public class ResultView {
     public static void print(int value) {
